Refactor: FONT_SET -> TEXT_SET으로 파일명 변경, TEXT_FORMAT 추가 정의 및 적용
