--- src/constants/FONT_SET.js
@@ -1,6 +1,0 @@
-export const FONTS = Object.freeze({
-  'Noto Sans': 'noto-sans',
-  Pretendard: 'pretendard',
-  나눔명조: 'nanum-gothic',
-  '나눔손글씨 손편지체': 'nanum-myeongjo',
-});
